Update integrations tests
