+++ app/views/posts/index.html.erb
@@ -8,7 +8,7 @@
 <div class="names_posts">
     <ul class="name_post"> 
    <li class="list_name"> <%= @user.name %> </li>
-   <li class="list_countpost">Number of post: <%= @user.postscounter %> </li>
+   <li class="list_countpost">Number of posts: <%= @user.postscounter %> </li>
    </ul>
 </div>
  </div>
