+++ spec/requests/posts_spec.rb
@@ -1,151 +1,43 @@
 require 'rails_helper'
 
-RSpec.describe 'posts', type: :feature do
-  describe 'posts/index' do
+RSpec.describe 'Posts', type: :request do
+  describe 'GET /users/id/posts/index' do
     before(:each) do
-      @user = [
-        User.create(
-          name: 'Chris',
-          photo: 'https://pics.com',
-          bio: 'Software Engineer Turkey',
-          postscounter: 2
-        )
-      ]
-
-      @first_user = User.first
-
-      @first_post = Post.create(
-        author: @first_user,
-        title: 'Hello',
-        text: 'This is my first post from Burak',
-        commentscounter: 1,
-        likescounter: 1
-      )
-      @second_post = Post.create(
-        author: @first_user,
-        title: 'How are you?',
-        text: 'This is my second post from Burak',
-        commentscounter: 0,
-        likescounter: 1
-      )
-      @first_comment = Comment.create(post: @first_post, author: @first_user, text: 'First comment for Burak')
-      @second_comment = Comment.create(post: @second_post, author: @first_user, text: 'Second comment for Burak')
-      @first_like = Like.create(post: @first_post, author: @first_user)
-      @second_like = Like.create(post: @second_post, author: @first_user)
-
-      visit user_posts_path(@first_user)
+      user = User.create!(name: 'Chris', photo: 'https://img.i.com/2.jpg', bio: 'Hey I am chris.',
+                          postscounter: 1)
+      get "/users/#{user.id}/posts"
     end
 
-    it 'shows the author image' do
-      expect(page).to have_css('img')
+    it 'check if it brings success response' do
+      expect(response).to be_successful
     end
 
-    it 'shows the author name' do
-      expect(page).to have_content(@first_user.name)
+    it 'renders the index template' do
+      expect(response).to render_template('index')
     end
 
-    it 'shows the number of posts by the author' do
-      expect(page).to have_content("Number of posts: #{@first_user.postscounter}")
-    end
-
-    it 'shows the post title' do
-      expect(page).to have_content(@first_post.title)
-    end
-
-    it 'shows the post text/body' do
-      expect(page).to have_content(@first_post.text)
-    end
-
-    it 'shows the first comments on the post' do
-      expect(page).to have_content(@first_comment.text)
-    end
-
-    it 'shows the number of comments on the post' do
-      expect(page).to have_content("Comments: #{@first_post.commentscounter}")
-    end
-
-    it 'shows the number of likes on the post' do
-      expect(page).to have_content("Likes: #{@first_post.likescounter}")
-    end
-
-    it 'shows a section for pagination if there are more posts than fit on the view' do
-      expect(page).to have_content('Pagination')
-    end
-
-    it "when click on a post, it redirects to the post's show page" do
-      visit user_posts_path(@first_user)
-      click_link @first_post.title
-      expect(current_path).to match(user_posts_path(@first_user.id))
+    it 'check the response body includes correct placeholder text' do
+      expect(response.body).to include('<ul class="photo"> ')
     end
   end
-  describe 'posts/show', type: :feature do
-    before(:each) do
-      @user = [
-        User.create(
-          name: 'Burak',
-          photo: 'https://pics.com',
-          bio: 'Software Engineer Turkey',
-          postscounter: 2
-        ),
-        User.create(
-          name: 'Mert',
-          photo: 'https://pics.com',
-          bio: 'Software Engineer from Turkey',
-          postscounter: 3
-        )
-      ]
 
-      @first_user = User.first
+  describe 'GET users from /users/:id to show' do
+    user = User.create!(name: 'Burak', photo: 'https://picsum.photos/200/300', bio: 'I am Burak', postscounter: 4)
+    post = Post.create!(author: user, title: 'title', text: ' Post text 1', commentscounter: 0, likescounter: 0)
 
-      @first_post = Post.create(
-        author: @first_user,
-        title: 'Hello',
-        text: 'This is my first post from Burak',
-        commentscounter: 1,
-        likescounter: 1
-      )
-
-      @first_comment = Comment.create(post: @first_post, author: @first_user, text: 'First comment for Burak')
-      @second_comment = Comment.create(post: @first_post, author: @first_user, text: 'Second comment for Burak')
-      @third_comment = Comment.create(post: @first_post, author: @first_user, text: 'Third comment for Burak')
-
-      @first_like = Like.create(post: @first_post, author: @first_user)
-      @second_like = Like.create(post: @first_post, author: @first_user)
-      @third_like = Like.create(post: @first_post, author: @first_user)
-
-      visit user_posts_path(@first_user, @first_post)
+    it 'check if it brings success response' do
+      get "/users/#{user.id}/posts/#{post.id}"
+      expect(response).to be_successful
     end
 
-    it 'shows posts title' do
-      expect(page).to have_content(@first_post.title)
+    it 'check the response body includes correct placeholder text' do
+      get "/users/#{user.id}/posts/#{post.id}"
+      expect(response.body).to include(' <div class="post_comment">')
     end
 
-    it 'shows posts author' do
-      expect(page).to have_content(@first_user.name)
-    end
-
-    it 'shows number of comments on the post' do
-      expect(page).to have_content(@first_post.commentscounter)
-    end
-
-    it 'shows number of likes on the post' do
-      expect(page).to have_content(@first_post.likescounter)
-    end
-
-    it 'shows the post text/body' do
-      expect(page).to have_content(@first_post.text)
-    end
-
-    it 'shows username of each commenters on the post' do
-      expect(page).to have_content(@first_comment.author.name)
-      expect(page).to have_content(@second_comment.author.name)
-      expect(page).to have_content(@third_comment.author.name)
-    end
-
-    it 'shows the text of each comment on the post' do
-      expect(page).to have_content(@first_comment.text)
-      expect(page).to have_content(@second_comment.text)
-      expect(page).to have_content(@third_comment.text)
+    it 'renders the show template' do
+      get "/users/#{user.id}/posts/#{post.id}"
+      expect(response).to render_template('show')
     end
   end
 end
